Local dev server uses bundled key/cert pair
Chrome 58 dropped support for CN validation in HTTPS.

The bundled webpack-dev-server doesn’t include SAN (the new hotness)

https://github.com/webpack/webpack-dev-server/issues/854

So the workaround of trusting the cert now fails.

This change adds our own cert that we can manage and have dev accept.

* It’s named “Mobify Local Development Server” so it is easy to find.
* It works on localhost and 127.0.0.1.
* It includes a SAN so it works with Chrome 58+

Also included is a script to regenerate it if we ever need to.